+++ src/controller/Modules/types/proposal.ts
@@ -1,5 +1,3 @@
-import { Contract, ethers } from 'ethers';
-import { VotesToken } from '../../../assets/typechain-types/module-governor';
 import { ProposalExecuteData } from '../../../types/proposal';
 
 export interface GovernanceProposalData {
@@ -7,21 +5,6 @@ export interface GovernanceProposalData {
     proposalData: ProposalExecuteData;
     successCallback?: () => void;
   }) => void;
-  governanceAddress?: string;
-  treasuryModuleContract?: Contract;
   pending?: boolean;
   isAuthorized?: boolean;
-  votingToken?: {
-    votingTokenContract: VotesToken | undefined;
-    votingTokenData: {
-      name: string | undefined;
-      symbol: string | undefined;
-      decimals: number | undefined;
-      userBalance: ethers.BigNumber | undefined;
-      userClaimAmount: ethers.BigNumber | undefined;
-      delegatee: string | undefined;
-      votingWeight: ethers.BigNumber | undefined;
-      address: string | undefined;
-    };
-  };
 }
